+++ fxn/types/prediction.py
@@ -3,7 +3,7 @@
 #   Copyright © 2024 NatML Inc. All Rights Reserved.
 #
 
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 from typing import Any, List, Optional
 
 from .predictor import PredictorType
@@ -17,33 +17,33 @@ class PredictionResource (BaseModel):
         type (str): Resource type.
         url (str): Resource URL.
     """
-    id: str
-    type: str
-    url: str
+    id: str = Field(description="Resource identifier.")
+    type: str = Field(description="Resource type.")
+    url: str = Field(description="Resource URL.")
 
 class Prediction (BaseModel):
     """
     Prediction.
 
     Members:
-        id (str): Prediction ID.
+        id (str): Prediction identifier.
         tag (str): Predictor tag.
         type (PredictorType): Prediction type.
         configuration (str): Prediction configuration token. This is only populated for `EDGE` predictions.
         resources (list): Prediction resources. This is only populated for `EDGE` predictions.
         results (list): Prediction results.
         latency (float): Prediction latency in milliseconds.
-        error (str): Prediction error. This is `null` if the prediction completed successfully.
+        error (str): Prediction error. This is `None` if the prediction completed successfully.
         logs (str): Prediction logs.
         created (str): Date created.
     """
-    id: str
-    tag: str
-    type: PredictorType
-    configuration: Optional[str] = None
-    resources: Optional[List[PredictionResource]] = None
-    results: Optional[List[Any]] = None
-    latency: Optional[float] = None
-    error: Optional[str] = None
-    logs: Optional[str] = None    
-    created: str
+    id: str = Field(description="Prediction identifier.")
+    tag: str = Field(description="Predictor tag.")
+    type: PredictorType = Field(description="Prediction type.")
+    configuration: Optional[str] = Field(default=None, description="Prediction configuration token. This is only populated for `EDGE` predictions.")
+    resources: Optional[List[PredictionResource]] = Field(default=None, description="Prediction resources. This is only populated for `EDGE` predictions.")
+    results: Optional[List[Any]] = Field(default=None, description="Prediction results.")
+    latency: Optional[float] = Field(default=None, description="Prediction latency in milliseconds.")
+    error: Optional[str] = Field(default=None, description="Prediction error. This is `None` if the prediction completed successfully.")
+    logs: Optional[str] = Field(default=None, description="Prediction logs.")
+    created: str = Field(description="Date created.")