+++ fxn/types/profile.py
@@ -3,7 +3,7 @@
 #   Copyright © 2024 NatML Inc. All Rights Reserved.
 #
 
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 from typing import Optional
 
 class Profile (BaseModel):
@@ -20,11 +20,11 @@ class Profile (BaseModel):
         website (str): User website.
         github (str): User GitHub handle.
     """
-    username: str
-    email: Optional[str] = None
-    created: Optional[str] = None
-    name: Optional[str] = None
-    avatar: Optional[str] = None
-    bio: Optional[str] = None
-    website: Optional[str] = None
-    github: Optional[str] = None
+    username: str = Field(description="Username.")
+    email: Optional[str] = Field(default=None, description="User email address.")
+    created: Optional[str] = Field(default=None, description="Date created.")
+    name: Optional[str] = Field(default=None, description="User display name.")
+    avatar: Optional[str] = Field(default=None, description="User avatar URL.")
+    bio: Optional[str] = Field(default=None, description="User bio.")
+    website: Optional[str] = Field(default=None, description="User website.")
+    github: Optional[str] = Field(default=None, description="User GitHub handle.")