+++ fxn/types/value.py
@@ -3,7 +3,7 @@
 #   Copyright © 2024 NatML Inc. All Rights Reserved.
 #
 
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 from typing import List, Optional, Union
 
 from .dtype import Dtype
@@ -14,9 +14,9 @@ class Value (BaseModel):
 
     Members:
         data (str): Value URL. This can be a web URL or a data URL.
-        type (Dtype): Value type.
+        type (Dtype): Value data type.
         shape (list): Value shape. This is `None` if shape information is not available or applicable.
     """
-    data: Union[str, None]
-    type: Dtype
-    shape: Optional[List[int]] = None
+    data: Union[str, None] = Field(description="Value URL. This can be a web URL or a data URL.")
+    type: Dtype = Field(description="Value data type.")
+    shape: Optional[List[int]] = Field(default=None, description="Value shape. This is `None` if shape information is not available or applicable.")